requirements.txt
pytesseract
pillow
pydantic
groq
python-dotenv
